Address PR feedback: Use float

+++ composeApp/src/commonMain/kotlin/com/daily/vitals/domain/entry/model/Entry.kt
@@ -9,5 +9,5 @@ data class Entry(
     val fasting: Int = 0,
     val postMeal: Int = 0,
     val sleep: Int = 0,
-    val weight: Double = 0.0
+    val weight: Float = 0f
 )
